refactor(freeze-updates): extract hook-queue pause/resume
Pull pauseHookQueue and resumeHookQueue out of freeze-updates.ts.
These wrap React's useState/useReducer/useSyncExternalStore queue
`pending` descriptor and `getSnapshot` function with a freeze-aware
shim, and on resume replay buffered actions through the original
dispatch.

The hook-queue module imports mergePendingChains/extractActionsFromChain
from the chain helpers module (extracted earlier), and the shared
state map + flag from freeze/state.js — no new shared variables
introduced.

freeze-updates.ts: 419 -> 345 (-74). 34 freeze specs green.
